refactor(macos): extract PlaylistTracks from AppModel into AppModel+PlaylistTracks.swift
